requirements.txt
db-dtypes==1.3.0
folium==0.14.0
google-cloud-bigquery==3.26.0
joblib==1.4.2
matplotlib==3.8.4
numpy==1.22.4
pandas==1.5.0
plotly==5.24.1
streamlit==1.39.0
streamlit-folium==0.23.0
urllib3==1.26.6